Fix bug where game ran with multiple grid states
- Remove initialisation of a 'grid' from the runGame method and remove
method parameters where they used the 'grid'.
Add second constructor so that board can still be passed in for test
purposes but in main game board instance variable is used as a default board (the old value
of the 'grid' parameter. This ensures only one state for the 'grid'.

- Remove grid parameter from hasAvailableMove() which lives on board so
 it should know the state of the grid and not be passed it.

- Remove grid parameter from Game.playMove() and Player.playMove().

- Remove 'grid' parameter from CommandLineUI getPlayerMove() and remove grid from
CommandLineUI tests.

- Rename isMoveUnique to isMoveAvailable.

- Rename getPlayerMove to getValidMove.